+++ src/models/issue.ts
@@ -5,12 +5,9 @@ const schema = Joi.object().keys({
   userHashId: Joi.string().allow(null).required().example('b45zo0'),
   assignedUserHashId: Joi.string().allow(null).required().example(null),
   title: Joi.string().required().example('Temperature is too high'),
-  pinGroupHashId: Joi.string().allow(null).required().example('dao97'), // TODO withthegrid/platform#1586 Rob: non-null in next release
+  pinGroupHashId: Joi.string().required().example('dao97'),
   level: Joi.number().integer().valid(0, 1, 2).required()
     .example(0),
-  typeKey: Joi.string().valid('missing', 'incorrect', 'unexpected', 'unrelated'), // TODO withthegrid/platform#1586 Rob: deprecated
-  startAt: Joi.date(), // TODO withthegrid/platform#1586 Rob: deprecated
-  endAt: Joi.date().allow(null).description('If null, the issue is still open'), // TODO withthegrid/platform#1586 Rob: deprecated
   closedAt: Joi.date().required().allow(null)
     .example(null)
     .description('If null, the issue is still open'),
@@ -27,11 +24,8 @@ interface Issue {
   userHashId: string | null;
   assignedUserHashId: string | null;
   title: string;
-  pinGroupHashId: string | null;
+  pinGroupHashId: string;
   level: 0 | 1 | 2;
-  typeKey?: 'missing' | 'incorrect' | 'unexpected' | 'unrelated';
-  startAt?: Date;
-  endAt?: Date | null;
   closedAt: Date | null;
   createdAt: Date;
   updatedAt: Date;