+++ src/routes/issue/add.ts
@@ -3,14 +3,10 @@ import { ControllerGeneratorOptionsWithClient } from '../../comms/controller';
 
 interface Request {
   body: {
-    links?: { // TODO withthegrid/platform#1586 Rob: deprecated
-      pinGroupHashId: string;
-      pinHashId: string | null;
-    }[];
     assignedUserHashId?: string | null;
-    pinGroupHashId?: string; // TODO withthegrid/platform#1586 Rob: required in future
-    pinHashIds?: string[]; // TODO withthegrid/platform#1586 Rob: required in future
-    automation?: { // TODO withthegrid/platform#1586 Rob: required in future
+    pinGroupHashId: string;
+    pinHashIds: string[];
+    automation: {
       type: 'missing'
     } | {
       type: 'thresholds'
@@ -18,12 +14,8 @@ interface Request {
     } | null;
     title: string;
     level: 0 | 1 | 2;
-    typeKey?: 'missing' | 'incorrect' | 'unexpected' | 'unrelated'; // TODO withthegrid/platform#1586 Rob: deprecated
     comment: string;
-    quantityHashIds?: string[]; // TODO withthegrid/platform#1586 Rob: deprecated
     labelHashIds?: string[];
-    startAt?: Date; // TODO withthegrid/platform#1586 Rob: deprecated
-    endAt?: Date | null; // TODO withthegrid/platform#1586 Rob: deprecated
   };
 }
 
@@ -40,14 +32,11 @@ const controllerGeneratorOptions: ControllerGeneratorOptionsWithClient = {
   method: 'post',
   path: '/',
   body: Joi.object().keys({
-    links: Joi.array().min(1).max(20).items(Joi.object().keys({
-      pinGroupHashId: Joi.string().required().example('dao97'),
-      pinHashId: Joi.string().allow(null).required().example(null),
-    })), // TODO withthegrid/platform#1586 Rob: deprecated
     assignedUserHashId: Joi.string().allow(null).default(null),
-    pinGroupHashId: Joi.string().example(['dao97']), // TODO withthegrid/platform#1586 Rob: required in future
-    pinHashIds: Joi.array().items(Joi.string()).example(['e13d57']).description('When empty, all pins are affected'), // TODO withthegrid/platform#1586 Rob: required in future
-    automation: Joi.alternatives().try( // TODO withthegrid/platform#1586 Rob: required in future
+    pinGroupHashId: Joi.string().required().example(['dao97']),
+    pinHashIds: Joi.array().items(Joi.string()).required().example(['e13d57'])
+      .description('When empty, all pins are affected'),
+    automation: Joi.alternatives().try(
       Joi.object({
         type: Joi.string().valid('missing').required(),
       }),
@@ -59,19 +48,15 @@ const controllerGeneratorOptions: ControllerGeneratorOptionsWithClient = {
           .required()
           .example(['sajia1']),
       }),
-    ).allow(null).example(null),
+    ).allow(null)
+      .required()
+      .example(null),
     title: Joi.string().max(100).required().example('Temperature is too high'),
     level: Joi.number().valid(0, 1, 2).required().example(0),
-    typeKey: Joi.string().valid('missing', 'incorrect', 'unexpected', 'unrelated'), // TODO withthegrid/platform#1586 Rob: deprecated
     comment: Joi.string().max(65536).allow('').required()
       .example('This looks serious.'),
-    // TODO withthegrid/platform#1586 Rob: deprecated
-    quantityHashIds: Joi.array().items(Joi.string()).max(10),
     labelHashIds: Joi.array().items(Joi.string()).max(10).default([])
       .example(['u98a24']),
-    startAt: Joi.date(), // TODO withthegrid/platform#1586 Rob: deprecated
-    endAt: Joi.date().allow(null), // TODO withthegrid/platform#1586 Rob: deprecated
-
   }).required(),
   right: { environment: 'ISSUES' },
   response: Joi.object().keys({