+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [17.73.1](https://github.com/withthegrid/platform-sdk/compare/v17.73.0...v17.73.1) (2022-12-23)
+
+
+### Bug Fixes
+
+* **measurement imports:** user experience enhancements ([#412](https://github.com/withthegrid/platform-sdk/issues/412)) ([888807f](https://github.com/withthegrid/platform-sdk/commit/888807fff0851610d382c82909838956d74cb09f)), closes [withthegrid/platform-client#1822](https://github.com/withthegrid/platform-client/issues/1822)
+
+## [17.73.0](https://github.com/withthegrid/platform-sdk/compare/v17.72.0...v17.73.0) (2022-12-21)
+
+
+### Features
+
+* **imports:** add import models and endpoints ([#378](https://github.com/withthegrid/platform-sdk/issues/378)) ([ef63c18](https://github.com/withthegrid/platform-sdk/commit/ef63c187ec9b25b935a1d942def0f47db62deafc))
+
 ## [17.72.0](https://github.com/withthegrid/platform-sdk/compare/v17.71.0...v17.72.0) (2022-12-02)